backend/requirements.txt
# SEBI Hackathon Stock Market Fraud Detection System
# Core Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Data Processing
pandas==2.1.0
numpy==1.25.2
python-dateutil==2.8.2
pytz==2023.3

# Financial Data
yfinance==0.2.28
yahoo-fin==0.8.9
investpy==1.0.8
alpha-vantage==2.3.1

# Web and API Interaction
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
httpx==0.25.2
tldextract==3.4.4

# Natural Language Processing and ML
scikit-learn==1.3.2
nltk==3.8.1
transformers==4.36.2
textblob==0.17.1
wordcloud==1.9.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Vector Database and RAG
chromadb==0.4.18
langchain==0.0.340
langchain-community==0.0.3
sentence-transformers==2.2.2

# LLM Integration
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2

# Authentication and Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4

# Caching and Performance
cachetools==5.3.1
redis==4.6.0

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
regex==2023.8.8
psutil==5.9.6
feedparser==6.0.10

# Social Media
newsapi-python==0.2.7
tweepy==4.14.0
praw==7.7.1
discord.py==2.3.2
python-telegram-bot==20.7
telethon==1.33.1

# Testing
pytest==7.4.0

# Concurrency and Async
asyncio==3.4.3
aiofiles==23.2.1
